部分功能修改

+++ src/main/java/com/qiton/model/MarkRecode.java
@@ -42,15 +42,15 @@ public class MarkRecode implements Serializable {
 
 	/** 收入 */
 	@TableField(value = "mrd_income")
-	private Float mrdIncome;
+	private Double mrdIncome;
 
 	/** 支出 */
 	@TableField(value = "mrd_pay")
-	private Float mrdPay;
+	private Double mrdPay;
 
 	/** 余钱 */
 	@TableField(value = "mrd_share")
-	private Float mrdShare;
+	private Double mrdShare;
 
 	/** 备注 */
 	@TableField(value = "mrd_remark")
@@ -62,7 +62,7 @@ public class MarkRecode implements Serializable {
 	}
 
 	
-	public MarkRecode(Long mrdUserid, String mrdUsername, Integer mrdProfittype, Date mrdTime, Float mrdShare,
+	public MarkRecode(Long mrdUserid, String mrdUsername, Integer mrdProfittype, Date mrdTime, Double mrdShare,
 			String mrdRemark) {
 		super();
 		this.mrdUserid = mrdUserid;
@@ -122,27 +122,27 @@ public class MarkRecode implements Serializable {
 		this.mrdTime = mrdTime;
 	}
 
-	public Float getMrdIncome() {
+	public Double getMrdIncome() {
 		return this.mrdIncome;
 	}
 
-	public void setMrdIncome(Float mrdIncome) {
+	public void setMrdIncome(Double mrdIncome) {
 		this.mrdIncome = mrdIncome;
 	}
 
-	public Float getMrdPay() {
+	public Double getMrdPay() {
 		return this.mrdPay;
 	}
 
-	public void setMrdPay(Float mrdPay) {
+	public void setMrdPay(Double mrdPay) {
 		this.mrdPay = mrdPay;
 	}
 
-	public Float getMrdShare() {
+	public Double getMrdShare() {
 		return this.mrdShare;
 	}
 
-	public void setMrdShare(Float mrdShare) {
+	public void setMrdShare(Double mrdShare) {
 		this.mrdShare = mrdShare;
 	}
